refactor(word): isolate word persistence transactions
